Warlock & Artificer (#634)
* Warlock

* Cloak of flies, pact weapon, knowledge changes

-turned pact weapon into a datum and allow user to change it's weapon type.
-changed cloak of flies to be damage over time
-knowledge also gives you the message spell

* Starchain change

-the chain only gives you devil's sight while equipped

* Boon changes

-removed revenge/curse
-added power/empowered eldritch blast

* Made talking weapons

Added singing_item.dm datum that can be added to all items to give them a random personality

Added helpful item_drop_

* Create singing_item.dm

* Pact weapon is possessable

You can channel the pact weapon to cause a ghost vote to possess it.

* Update roguetown.dme

* ghost poll works

* channel on middle click instead of attack_self

* soulbond ring

* after soullink dies wait 30 seconds before killing soulsharer

* Find Familiar

* Find Familiar Update

* force faithless

* Fiend Patron

* Soulbond ring no longer dusts

* Celestial can heal non-faithless too

* Artificer Infusions

Artificer is a trait. No player starts with it. A player with artificer can infuse items with magical abilities. This is not balanced yet and should not be given to players until I limit the amount of infused items they can equip.

* of the sewers and haunting

haunting and of the sewers instruments now work

* sending stones