Final Push for the day (map-stuff)
adding some touches to the undercity such as a place to buy drugs, and bolted doors they might prefer for their privacy.

created a vault for the guild with the steward role. (Will be turned into an Appraiserfor the Guild)

added some improvements for the dwarf keep to help with the adventurer tide as a dwarf artificer.