Malicious Mob-targetting
trying to help the server with the latency and lag led me to realizing a lot of spawners existed for haunts before event procs. I have removed as many as I could find, as well as some other mobs that were mostly littering up areas or not really contributing to the round.

This a bandage towards the longterm fix proposed by stonedelf to create mob spawners which have people around, as well as work on the ambush tables with area-zones. Both are wanted to create more of a scaled content and difficulty associated to places on the pre-existing map in line with high fantasy themes. (Mythology, Tabletop Games, Fantasy)