Excidium removal
The map is not intended to have an Excidium in the city. This interferes with the lawless/lawful themes of the server and mechanically gives them a reason to kill players with the Outlaw/Criminal perk.

What is instead intended is people use this to indicate character themes and the repercussion is handled ICly.