Changed the excidium to be only in the ranger-sentinels base.
People will only ever use the excidium on those beyond redemption/RP at this point. It is extremely discouraged to have a mechanical aspect to bounty-hunting and may be entirely removed in favor making this gameplay IC/RP based. Don't be suprised if it's gone forever if the issue persists server-side. I hate excidium.