Making some light Magick (#117)
* Making some light Magick

This is a very small PR. It adds some light items of food to the guild kitchen, as well as some cheese to the mage academy basement. Beyond this the main aspect is changing the iron bars around the magic school to permanent magick forcefields on the windows! (looks much nicer. Tested it in game!) -- this was also given to the church windows. Because the luneites and magicians are quite good pals ICly!

- Added some statues to the upper part of the lune church and also worked on the lower windows that had an issue being underwater.

* Armory lever fix

- Armory Lever HOTFIX