General Sweeping and a bit of Wild Magick (Revised)
Loadout changes for: druid, hedgeknight/Hedgemaster (giving them the sylvan stuff and otherwise enhancing whats already there. Druids have their seedbank nerfed in exchange for lifting them out of the muck and into a role that can opt to bless crops but exists moreso for forest spirituality and law matters.

Increased the PQ for mortician to 15. I understand it was originally intended to 'redeem' PQ deficits but not any more.

Monk was removed because the role doesn't even do what people think it does.

probability and job slots have been increased to the other dm's touched.