Particle Weather System & Food smells (#1733)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Ports:
- https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/pull/18.
- https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/pull/20.
- https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/pull/21.

The weather types include:
- Rain
- Rainstorm
- Snow
- Snowstorm

Testing proof:

https://github.com/user-attachments/assets/bd7108c4-4d0a-4d06-9760-c8e5fa76f685

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Functioning weather after a long time.
Food smells make use of the pollutant system that was only used for
perfumes.

---------

Co-authored-by: dwasint <82520990+dwasint@users.noreply.github.com>
# Conflicts:
#	code/_onclick/other_mobs.dm
#	code/modules/food_and_drinks/food/snacks.dm
#	code/modules/mob/living/carbon/life.dm
#	modular/Neu_Food/code/raw/NeuFood_veggies.dm
#	roguetown.dme