Fixes issue with helmets/masks not allowing you to drink water. (#1785)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changed the proc it was calling for one that properly checks if you are
wearing something that covers your mouth.


https://github.com/user-attachments/assets/4a626d04-7eb4-4026-9c66-2f81a206a5f6


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
fixes #534
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->