The Mogging Pt. ??
PQ Adjustments, gear Adjustments and Mapping (Oh my!)

The PQ adjustments represent the RP requirement I have on entry level roles; so that people have something to strive for. Mercenary, shaman, chieftain and sellsword will be receiving PQ locks because they require competent leaders and in the mercenaries case, because they will be getting a boost alongside new loadouts like a arquebus wielding rifleman to give mercs an edge vs adventurers.

Much requested upgrade of the guild hall. (Whole new level!)
Cafeteria to cause them woe as they realize the only way in to the kitchen is now the second story.

shaman had his loadout tweaked, and I am slowly starting to focus on the tribal clan and expand it further into the woods.

There is a new way to access the underdark for people to find, making the dwarfkeep easier to rush, or arguably less dangerous to access if they are smart.