Update objectives_rogue.dm
This needs reviewing, just in case. These are all name/lore changes that should serve to help with worldbuilding for antags. In most cases I've shifted away from outright telling them to kill players in favor of finding their own style of RP.

Lore is shifting to a place where there are more concepts and ambiguity vs entrappings.

In both cases of the w/w and vamp the goal is to 'broaden your clan' and 'survive' with no actual declaration to attack or outright siege the city. This is something that should happen lateround if at all.