Videdev (#255)
* Helmets no longer hide hair if they are in your belt. (#1780)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes the proc that hides hair to check if the item is in the belt slot,
doing this it won't remove the hair overlay when is not in the head
slot.


https://github.com/user-attachments/assets/b831bdb3-13c7-4124-902d-91ae1d83b51d


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It fixes a bug.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

* Fixes issue with helmets/masks not allowing you to drink water. (#1785)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changed the proc it was calling for one that properly checks if you are
wearing something that covers your mouth.


https://github.com/user-attachments/assets/4a626d04-7eb4-4026-9c66-2f81a206a5f6


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
fixes #534
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

* stuff

* helmet and 10 pq 1 extra quirk point

* balloon

* Update roguetown.dme

* fix scroll fuckery probably

* Update granters.dm

* migrussy

---------

Co-authored-by: TheFenix323 <64550272+TheFenix323@users.noreply.github.com>